A detour, a couple defeats, and a marginal victory
printing colors from a function, but now that its done, I realize I want to use impl and try and monkey patch strings or numbers to be able
to print their color, or something wacky like that. But thats probably
too hard for me right now. Or no probable in rust….or not!